+++ CHANGELOG.md
@@ -8,8 +8,8 @@ New features and improvements:
 
 - Added ``normalize()`` method from shapely to GeoSeries/GeoDataframe (#2537)
 - Add where filter to ``read_file`` (#2552)
-- ``explore()`` enhanced to work with more combinations of 'cmap', 'scheme', 'legend'
-  and 'category' arguments (#2590)
+- ``explore()`` enhanced to work with more combinations of 'column', 'cmap', 
+  'color', 'legend', 'categorical', 'scheme', 'k', 'vmin' and 'vmax' arguments (#2590)
 
 Deprecations and compatibility notes: